added JsonOptions to AddControllers method
added JsonOption because of an error that ocurred when tried to get all pets entities with their owners which caused the error (System.Text.Json.JsonException: A possible object cycle was detected. This can either be due to a cycle or if the object depth is larger than the maximum allowed depth of 32. Consider using ReferenceHandler.Preserve on JsonSerializerOptions to support cycles). And after adding that option to the AddControllers method it FIXED the problem